+++ uncertainties/core/formatting.py
@@ -1,5 +1,11 @@
 
 import math
+from math import sqrt, log, isnan, isinf  # Optimization: no attribute look-up
+try:
+    from math import isinfinite  # !! Python 3.2+
+except ImportError:
+    def isinfinite(x):
+        return isinf(x) or isnan(x)
 
 def first_digit(value):
     '''
