+++ uncertainties/test_uncertainties.py
@@ -30,6 +30,7 @@ import sys
 # Local modules
 
 import uncertainties.core.core as uncert_core
+import uncertainties.core.formatting as uncert_format
 from uncertainties.core.core import ufloat, AffineScalarFunc, ufloat_fromstr
 from uncertainties import umath
 
@@ -2138,21 +2139,21 @@ def test_custom_pretty_print_and_latex():
 
     # We will later restore the defaults:
     PREV_CUSTOMIZATIONS = {
-        var: getattr(uncert_core, var).copy()
+        var: getattr(uncert_format, var).copy()
         for var in ['PM_SYMBOLS', 'MULT_SYMBOLS', 'GROUP_SYMBOLS']}
     
     # Customizations:
     for format in ["pretty-print", "latex"]:
-        uncert_core.PM_SYMBOLS[format] = u" ± "
-        uncert_core.MULT_SYMBOLS[format] = u"⋅"
-        uncert_core.GROUP_SYMBOLS[format] = ( "[", "]" )
+        uncert_format.PM_SYMBOLS[format] = u" ± "
+        uncert_format.MULT_SYMBOLS[format] = u"⋅"
+        uncert_format.GROUP_SYMBOLS[format] = ( "[", "]" )
 
     assert u"{:P}".format(x) == u'[2.00 ± 0.10]⋅10⁻¹¹'
     assert u"{:L}".format(x) == u'[2.00 ± 0.10] ⋅ 10^{-11}'
 
     # We restore the defaults:
     for (var, setting) in PREV_CUSTOMIZATIONS.items():
-        setattr(uncert_core, var, setting)
+        setattr(uncert_format, var, setting)
 
 ###############################################################################
